[SW-1975] Move spark.ext.h2o.hdfs_conf Property among Properties of Internal Backend
